+++ contrib/install-osvvm.tcl
@@ -5,5 +5,5 @@ set ::osvvm::AnalyzeErrorStopCount 1
 
 foreach std {2008 2019} {
     SetVHDLVersion $std
-    source "OsvvmLibraries.pro"
+    build OsvvmLibraries
 }